Deduplicate PostHog ops helpers
